buy funtionality

+++ Monopoly/Tiles/HotelBuilt.cs
@@ -16,7 +16,7 @@ namespace Monopoly
             this.tile = tile;
         }
 
-        public double getPrice()
+        public int getPrice()
         {
             return this.tile.getPrice() * 2;
         }
